+++ src/Tests.ToolKit/Data/Lite/LiteDbRepositorySpecs/FilterLiteDbRepositoryTests.cs
@@ -2,6 +2,7 @@
 using FakeItEasy;
 using FatCat.Fakes;
 using FatCat.Toolkit.Testing;
+using FluentAssertions;
 
 namespace Tests.FatCat.Toolkit.Data.Lite.LiteDbRepositorySpecs;
 
@@ -20,5 +21,29 @@ public abstract class FilterLiteDbRepositoryTests<T> : RequireCollectionLiteDbRe
 
 		A.CallTo(() => collection.Find(filterCapture, A<int>._, A<int>._))
 		.Returns(ItemsToReturn);
+	}
+
+	protected void SetUpFindWithEmptyCollection()
+	{
+		A.CallTo(() => collection.Find(filterCapture, A<int>._, A<int>._))
+		.Returns(new List<LiteDbTestObject>());
+	}
+
+	protected void VerifyFilterCallOnCollectionMade()
+	{
+		A.CallTo(() => collection.Find(filterCapture, A<int>._, A<int>._))
+		.MustHaveHappened();
+
+		var expression = filterCapture.Value.Compile();
+
+		var filterItem = new LiteDbTestObject { SomeNumber = numberToFind };
+
+		expression(filterItem)
+			.Should()
+			.BeTrue();
+
+		expression(new LiteDbTestObject { SomeNumber = numberToFind - 1 })
+			.Should()
+			.BeFalse();
 	}
 }
