+++ src/Tests.ToolKit/Data/Lite/LiteDbRepositorySpecs/GetAllByFilterTests.cs
@@ -14,20 +14,7 @@ public class GetAllByFilterTests : FilterLiteDbRepositoryTests<List<LiteDbTestOb
 	{
 		await RunTest();
 
-		A.CallTo(() => collection.Find(filterCapture, A<int>._, A<int>._))
-		.MustHaveHappened();
-
-		var expression = filterCapture.Value.Compile();
-
-		var filterItem = new LiteDbTestObject { SomeNumber = numberToFind };
-
-		expression(filterItem)
-			.Should()
-			.BeTrue();
-
-		expression(new LiteDbTestObject { SomeNumber = numberToFind - 1 })
-			.Should()
-			.BeFalse();
+		VerifyFilterCallOnCollectionMade();
 	}
 
 	[Fact]
@@ -39,4 +26,6 @@ public class GetAllByFilterTests : FilterLiteDbRepositoryTests<List<LiteDbTestOb
 	}
 
 	protected override Task<List<LiteDbTestObject>> RunTest() => repository.GetAllByFilter(i => i.SomeNumber == numberToFind);
+
+	
 }
