+++ src/Tests.ToolKit/Data/Lite/LiteDbRepositorySpecs/GetByFilterTests.cs
@@ -16,25 +16,13 @@ public class GetByFilterTests : FilterLiteDbRepositoryTests<LiteDbTestObject>
 	{
 		await RunTest();
 
-		A.CallTo(() => collection.Find(filterCapture, A<int>._, A<int>._))
-		.MustHaveHappened();
-
-		var expression = filterCapture.Value.Compile();
-
-		expression(new LiteDbTestObject { SomeNumber = numberToFind })
-			.Should()
-			.BeTrue();
-
-		expression(new LiteDbTestObject { SomeNumber = numberToFind - 1 })
-			.Should()
-			.BeFalse();
+		VerifyFilterCallOnCollectionMade();
 	}
 
 	[Fact]
 	public async Task IfNothingFoundReturnNull()
 	{
-		A.CallTo(() => collection.Find(filterCapture, A<int>._, A<int>._))
-		.Returns(new List<LiteDbTestObject>());
+		SetUpFindWithEmptyCollection();
 
 		var result = await RunTest();
 
@@ -52,4 +40,6 @@ public class GetByFilterTests : FilterLiteDbRepositoryTests<LiteDbTestObject>
 	}
 
 	protected override Task<LiteDbTestObject> RunTest() => repository.GetByFilter(i => i.SomeNumber == numberToFind);
+
+	
 }
